Separate non-ROS-dependent YAML import functions
